I added an argument "part". There are some ELAN files in which we have to mark distinct sessions. This has very much potential, especially when we take into account all sorts of linked relations which are very common in our data. Naturally, this "part" works only with files where this field is adequately filled. But in principle it works like this:
kpv_vanejev <- read_eaf(eaf.list = "./kpv_udora/kpv_udo19420000SFOuVanejevMN.eaf", part = "partT")

In this case the part tier is filled so that each utterance has corresponding annotation on an independent part-tier. The better solution would be to fill it like this:

story
-
-
-
-
song
-
-
-
unmarked
-
-
-

And we can use the package 'zoo', as an example, to fill missing values with the last known one.

I removed the nicer sorting of columns from the end, this has to be thought about. Filename should not be the first one as it takes all the space on the screen. But how to order them without listing their names?